Move ProposalBuffer to its own file.
