+++ validator/src/lib.rs
@@ -5,5 +5,6 @@ pub mod aggregation;
 mod jail;
 mod r#macro;
 mod micro;
+mod proposal_buffer;
 pub mod tendermint;
 pub mod validator;
