+++ validator/src/validator.rs
@@ -6,12 +6,11 @@ use std::{
         atomic::{AtomicBool, Ordering},
         Arc,
     },
-    task::{Context, Poll, Waker},
+    task::{Context, Poll},
     time::Duration,
 };
 
-use futures::stream::{BoxStream, Stream, StreamExt};
-use linked_hash_map::LinkedHashMap;
+use futures::stream::{BoxStream, StreamExt};
 use nimiq_block::{Block, BlockHeaderTopic, BlockTopic, BlockType, EquivocationProof};
 use nimiq_blockchain::{BlockProducer, Blockchain};
 use nimiq_blockchain_interface::{AbstractBlockchain, BlockchainEvent, ForkEvent, PushResult};
@@ -22,15 +21,13 @@ use nimiq_database::{
     DatabaseProxy, TableProxy,
 };
 use nimiq_hash::{Blake2bHash, Hash};
-use nimiq_keys::{Address, KeyPair as SchnorrKeyPair, Signature as SchnorrSignature};
-use nimiq_macros::store_waker;
+use nimiq_keys::{Address, KeyPair as SchnorrKeyPair};
 use nimiq_mempool::{config::MempoolConfig, mempool::Mempool};
 use nimiq_network_interface::{
-    network::{MsgAcceptance, Network, NetworkEvent, PubsubId, SubscribeEvents, Topic},
+    network::{MsgAcceptance, Network, NetworkEvent, SubscribeEvents},
     request::request_handler,
 };
 use nimiq_primitives::{coin::Coin, policy::Policy};
-use nimiq_tendermint::SignedProposalMessage;
 use nimiq_transaction_builder::TransactionBuilder;
 use nimiq_validator_network::ValidatorNetwork;
 use parking_lot::RwLock;
@@ -39,12 +36,10 @@ use tokio_metrics::TaskMonitor;
 use tokio_stream::wrappers::BroadcastStream;
 
 use crate::{
-    aggregation::tendermint::{
-        proposal::{Header, RequestProposal},
-        state::MacroState,
-    },
+    aggregation::tendermint::{proposal::RequestProposal, state::MacroState},
     jail::EquivocationProofPool,
     micro::{ProduceMicroBlock, ProduceMicroBlockEvent},
+    proposal_buffer::{ProposalBuffer, ProposalReceiver},
     r#macro::{MappedReturn, ProduceMacroBlock, ProposalTopic},
 };
 
@@ -918,107 +913,5 @@ where
         }
 
         Poll::Pending
-    }
-}
-
-type ProposalAndPubsubId<TValidatorNetwork> = (
-    <ProposalTopic<TValidatorNetwork> as Topic>::Item,
-    <TValidatorNetwork as ValidatorNetwork>::PubsubId,
-);
-
-struct ProposalBuffer<TValidatorNetwork: ValidatorNetwork + 'static>
-where
-    <TValidatorNetwork as ValidatorNetwork>::PubsubId: std::fmt::Debug + Unpin,
-{
-    buffer: LinkedHashMap<
-        <TValidatorNetwork::NetworkType as Network>::PeerId,
-        ProposalAndPubsubId<TValidatorNetwork>,
-    >,
-    waker: Option<Waker>,
-}
-
-impl<TValidatorNetwork: ValidatorNetwork + 'static> ProposalBuffer<TValidatorNetwork>
-where
-    <TValidatorNetwork as ValidatorNetwork>::PubsubId: std::fmt::Debug + Unpin,
-{
-    // Ignoring clippy warning: this return type is on purpose
-    #[allow(clippy::new_ret_no_self)]
-    pub fn new() -> (
-        ProposalSender<TValidatorNetwork>,
-        ProposalReceiver<TValidatorNetwork>,
-    ) {
-        let buffer = Self {
-            buffer: LinkedHashMap::new(),
-            waker: None,
-        };
-        let shared = Arc::new(RwLock::new(buffer));
-        let sender = ProposalSender {
-            shared: Arc::clone(&shared),
-        };
-        let receiver = ProposalReceiver { shared };
-        (sender, receiver)
-    }
-}
-
-struct ProposalSender<TValidatorNetwork: ValidatorNetwork + 'static>
-where
-    <TValidatorNetwork as ValidatorNetwork>::PubsubId: std::fmt::Debug + Unpin,
-{
-    shared: Arc<RwLock<ProposalBuffer<TValidatorNetwork>>>,
-}
-
-impl<TValidatorNetwork: ValidatorNetwork + 'static> ProposalSender<TValidatorNetwork>
-where
-    <TValidatorNetwork as ValidatorNetwork>::PubsubId: std::fmt::Debug + Unpin,
-{
-    pub fn send(&self, proposal: ProposalAndPubsubId<TValidatorNetwork>) {
-        let source = proposal.1.propagation_source();
-        let mut shared = self.shared.write();
-        shared.buffer.insert(source, proposal);
-        if let Some(waker) = shared.waker.take() {
-            waker.wake()
-        }
-    }
-}
-
-struct ProposalReceiver<TValidatorNetwork: ValidatorNetwork + 'static>
-where
-    <TValidatorNetwork as ValidatorNetwork>::PubsubId: std::fmt::Debug + Unpin,
-{
-    shared: Arc<RwLock<ProposalBuffer<TValidatorNetwork>>>,
-}
-
-impl<TValidatorNetwork: ValidatorNetwork + 'static> Stream for ProposalReceiver<TValidatorNetwork>
-where
-    <TValidatorNetwork as ValidatorNetwork>::PubsubId: std::fmt::Debug + Unpin,
-{
-    type Item = SignedProposalMessage<
-        Header<<TValidatorNetwork as ValidatorNetwork>::PubsubId>,
-        (SchnorrSignature, u16),
-    >;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let mut shared = self.shared.write();
-        if shared.buffer.is_empty() {
-            store_waker!(shared, waker, cx);
-            Poll::Pending
-        } else {
-            let value = shared.buffer.pop_front().map(|entry| {
-                let (msg, id) = entry.1;
-                msg.into_tendermint_signed_message(Some(id))
-            });
-            Poll::Ready(value)
-        }
-    }
-}
-
-impl<TValidatorNetwork: ValidatorNetwork + 'static> Clone for ProposalReceiver<TValidatorNetwork>
-where
-    <TValidatorNetwork as ValidatorNetwork>::PubsubId: std::fmt::Debug + Unpin,
-{
-    fn clone(&self) -> Self {
-        Self {
-            shared: Arc::clone(&self.shared),
-        }
     }
 }
